Wrap Google Admin SDK in an auth flow
Introduces DCI library and the first context (GetOauthUser)
Implements a basic flow for configuring and regular auth via Google apps

TODO
The Google Admin API call is not finished - merely wrapped inside a
flow. Next steps are to extract and structure it properly, then start
making use of the data we can access.